backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==1.0.3
python-dateutil==2.8.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0